Corregido Filtro al momento de buscar los productos del stock desde la web

+++ workspace/controlador/src/main/java/ec/com/codesoft/codefaclite/controlador/inventario/ReporteInventarioStockControlador.java
@@ -149,7 +149,7 @@ public class ReporteInventarioStockControlador extends ModelControladorAbstract<
             
             for (Object[] objeto : listaStock) {
                 Producto producto = (Producto) objeto[0];
-                Integer cantidad = (Integer) objeto[1];
+                BigDecimal cantidad = (BigDecimal) objeto[1];
                 BigDecimal costoPromedio = (BigDecimal) objeto[2];
                 //Kardex kardexTemp = (Kardex) objeto[2];
